# api/openapi.yaml
openapi: 3.0.3
info:
  title: Flywheel API
  description: Local operator REST API for Flywheel projects, tickets, reviews, and workflows. No login or bearer token is required; the server binds to loopback and enforces local Host and same-origin browser requests. MCP authenticates separately with agent credentials.
  version: 0.1.0
servers:
  - url: http://localhost:8090
    description: Local
paths:
...
  /projects/{projectID}:
    get:
      operationId: GetProject
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
    patch:
      operationId: UpdateProject
      summary: Update project (e.g. set status to active or closed)
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        "400":
          description: Bad request (e.g. invalid status)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
...
  /tickets/{ticketID}:
    get:
      operationId: GetTicket
      parameters:
        - name: ticketID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
    patch:
      operationId: UpdateTicket
      parameters:
        - name: ticketID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTicketRequest'
      responses:
        "204":
          description: No content
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
...
  /tickets/{ticketID}/transitions:
    get:
      operationId: GetTransitions
      summary: State transition history for a ticket (timeline of state changes)
      parameters:
        - name: ticketID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransitionHistory'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
    post:
      operationId: TransitionTicket
      summary: Transition ticket state (e.g. start, submit, approve, reject)
      parameters:
        - name: ticketID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransitionRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad request (e.g. trigger required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
...
  /api/v1/gate/callback/{token}:
    post:
      operationId: PostGateCallback
      summary: Receive a webhook callback that satisfies a gate condition
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Gate condition marked as satisfied
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  ticket_id:
                    type: string
                  phase_id:
                    type: string
        "404":
          description: Token not found or expired
...
components:
  schemas:
...
    StructuredError:
      type: object
      required: [error, code, retriable]
      properties:
        error:
          type: string
        code:
          type: string
          enum: [lease_expired, unauthorized, not_found, conflict, invalid_input, forbidden, internal, not_implemented, project_closed]
        retriable:
          type: boolean
...
    TransitionRequest:
      type: object
      required: [trigger]
      properties:
        trigger:
          type: string
        actor_id:
          type: string
        actor:
          type: string
          enum: [human, agent, system]
        payload:
          type: object
...
    Project:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          description: Short project description.
        repo_url:
          type: string
        default_branch:
          type: string
          description: Branch to checkout when closing a work stream; default "main".
        tech_stack:
          type: array
          items:
            type: string
        context_pack:
          type: object
        status:
          type: string
          enum: [active, closed]
          description: active (default) or closed; list endpoints default to active only.
        dispatch_enabled:
          type: boolean
          description: Whether the dispatcher picks up tickets for this project (default true).
        dispatch_config:
          $ref: '#/components/schemas/DispatchConfig'
        created_at:
          type: string
          format: date-time
    UpdateProjectRequest:
      type: object
      properties:
        status:
          type: string
          enum: [active, closed]
        name:
          type: string
          description: Project display name.
        description:
          type: string
          description: Short project description.
        slug:
          type: string
          description: URL-safe identifier (unique per org).
        repo_url:
          type: string
          description: Git repo URL; when set, enables work streams + git integration (branch instructions in MCP).
        default_branch:
          type: string
          description: Branch to checkout when closing a work stream; default "main".
        dispatch_enabled:
          type: boolean
          description: Whether the dispatcher picks up tickets for this project (default true). Takes effect immediately.
        dispatch_config:
          $ref: '#/components/schemas/DispatchConfig'
        context_pack:
          type: object
          description: Project context pack (system_prompt, conventions, key_files).
          properties:
            system_prompt:
              type: string
            conventions:
              type: string
            key_files:
              type: array
              items:
                type: object
                properties:
                  path:
                    type: string
                  snippet:
                    type: string
    DispatchConfig:
      type: object
      properties:
        max_active_workers:
          type: integer
          minimum: 0
          description: Maximum concurrent active workers for this project. Set 0 or omit to inherit the server DISPATCH_MAX_WORKERS default.
        roles:
          type: array
          description: Custom dispatch roles. Each role has its own routing policy and inherits prompt/tool behavior from a built-in worker type.
          items:
            $ref: '#/components/schemas/DispatchWorkerRole'
        workers:
          type: array
          items:
            $ref: '#/components/schemas/DispatchWorkerProfile'
        policies:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/DispatchRolePolicy'
...
    DispatchRolePolicy:
      type: object
      properties:
        selection_mode:
          type: string
          enum: [ordered, any]
          description: ordered tries workers in the configured order; any rotates across the configured workers.
        worker_ids:
          type: array
          description: Worker profile IDs to use for this role. Leave empty to use all enabled workers.
          items:
            type: string
...
    Objective:
      type: object
      properties:
        description:
          type: string
        success_criteria:
          type: array
          items:
            type: string
        acceptance_test:
          type: string
    TicketContext:
      type: object
      properties:
        relevant_files:
          type: array
          items:
            type: string
        constraints:
          type: array
          items:
            type: string
        prior_attempts:
          type: array
          items:
            type: object
        human_answers:
          type: array
          items:
            type: string
    Ticket:
      type: object
      properties:
        id:
          type: string
        project_id:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [task, bug, spike, review]
        priority:
          type: integer
          minimum: 0
          maximum: 3
        state:
          type: string
          enum: [draft, planning, awaiting_input, executing, awaiting_validation, validated, closed]
        version:
          type: integer
        objective:
          $ref: '#/components/schemas/Objective'
        ticket_context:
          $ref: '#/components/schemas/TicketContext'
        inputs:
          type: object
        outputs:
          type: object
        depends_on:
          type: array
          items:
            type: string
        work_stream_id:
          type: string
        environment_id:
          type: string
          description: Compound environment ID (spec v0.2 section 4.1)
        target_repo:
          type: string
          description: Repository alias for multi-repo projects (from project_repositories). Empty means primary repo.
        assigned_to:
          type: string
        workflow_id:
          type: string
          description: Active workflow definition ID. Empty means legacy state machine.
        workflow_phase:
          type: string
          description: Current phase ID within the workflow. Empty means no active phase.
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        external:
          $ref: '#/components/schemas/TicketExternalRef'
...
    UpdateTicketRequest:
      type: object
      properties:
        depends_on:
          type: array
          items:
            type: string
        title:
          type: string
          description: When set, replaces the ticket title.
        objective:
          $ref: '#/components/schemas/Objective'
          description: Partial update; only fields present in the body are merged into the existing objective. Task/bug tickets must still have at least one of success_criteria or acceptance_test after the merge.
...
    TransitionHistory:
      type: object
      properties:
        ticket_id:
          type: string
        current_state:
          type: string
        transitions:
          type: array
          items:
            $ref: '#/components/schemas/StateTransitionEntry'
    StateTransitionEntry:
      type: object
      properties:
        id:
          type: string
        from_state:
          type: string
        to_state:
          type: string
        trigger:
          type: string
        actor_id:
          type: string
        actor_type:
          type: string
          enum: [human, agent, system]
        created_at:
          type: string
          format: date-time
    # --- Plan schemas ---
    # --- Policy calibration schemas ---
...